+++ Sources/fido2/PublicKeyCredential.swift
@@ -194,7 +194,6 @@ public struct PublicKeyCredentialUserEntity: Codable {
     }
 }
 
-
 /// Represents additional parameters when creating a new credential.
 public struct PublicKeyCredentialParameters: Codable {
     /// The type of credential to be created.
@@ -226,7 +225,6 @@ public struct PublicKeyCredentialParameters: Codable {
         alg = COSEAlgorithmIdentifier.parse(from: value)
     }
 }
-
 
 /// Represents credential parameters to be used for FIDO2 registration or authentication.
 public struct PublicKeyCredentialDescriptor: Codable {
@@ -430,6 +428,12 @@ public enum AuthenticatorTransport: String, Codable {
     
     /// Authenticator can be contacted over Bluetooth Low Energy (BLE).
     case ble
+    
+    /// Authenticator can be contacted over ISO/IEC 7816 smart card with contacts.
+    case smartcard = "smart-card"
+    
+    /// Authenticator can be contacted using a combination of (often separate) data-transport and proximity mechanisms. This supports, for example, authentication on a desktop computer using a smartphone.
+    case hybrid
     
     /// Authenticator is contacted using a client device-specific transport, i.e a platform authenticator. These authenticators are not removable from the client device.
     case `internal` = "internal"